refactor: derive color tokens from shared palette
